requirements.txt
altgraph==0.17.3
attrs==22.2.0
auto-py-to-exe==2.28.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Eel==0.14.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
isodate==0.6.1
lxml==4.9.2
numpy==1.24.1
onvif-zeep==0.2.12
opencv-python==4.7.0.68
pefile==2022.5.30
platformdirs==2.6.2
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.14
whichcraft==0.6.1
zeep==4.2.1
zope.event==4.6
zope.interface==5.5.2
